feat:세션 라우팅 설정 및 상세페이지 생성:YPW-108
