+++ src/pages/admin/sessions/SessionEdit.tsx
@@ -1,0 +1,7 @@
+import { FC } from 'react';
+
+const SessionEdit: FC = () => {
+  return <div></div>;
+};
+
+export default SessionEdit;
