+++ src/pages/admin/sessions/SessionWrite.tsx
@@ -1,0 +1,6 @@
+import { FC } from 'react';
+
+const SessionWrite: FC = () => {
+  return <div></div>;
+};
+export default SessionWrite;
