+++ src/routes/RouteSetup.tsx
@@ -9,7 +9,9 @@ import MemberList from '@pages/admin/members/MemberList';
 import Notice from '@pages/admin/notices/Notice.tsx';
 import NoticeList from '@pages/admin/notices/NoticeList';
 import NoticeWrite from '@pages/admin/notices/NoticeWrite';
-import Session from '@pages/admin/sessions/Session.tsx';
+import Session from '@pages/admin/sessions/Session';
+import SessionList from '@pages/admin/sessions/SessionList';
+import SessionWrite from '@pages/admin/sessions/SessionWrite';
 import SettingLink from '@pages/admin/settings/SettingLink';
 import SettingUpdate from '@pages/admin/settings/SettingUpdate';
 import AuthGuard from '@pages/auth/AuthGuard';
@@ -53,9 +55,9 @@ export const RouteSetup = () => {
           </Route>
 
           <Route path="sessions">
-            <Route index element={<Session />} />
-            <Route element={<NoticeWrite />} path="write" />
-            <Route element={<Notice />} path="detail/:id" />
+            <Route index element={<SessionList />} />
+            <Route element={<SessionWrite />} path="write" />
+            <Route element={<Session />} path="detail/:id" />
           </Route>
 
           <Route path="settings">
