+++ src/utils/getSessionType.ts
@@ -7,7 +7,7 @@ interface SessionReturnType {
   style: ChipStyle;
 }
 
-export const getSessionType = (type: SessionType): SessionReturnType => {
+export const getSessionType = (type?: SessionType): SessionReturnType => {
   switch (type) {
     case 'OFFLINE':
       return { text: '오프라인', color: 'primary', style: 'weak' };
@@ -15,5 +15,7 @@ export const getSessionType = (type: SessionType): SessionReturnType => {
       return { text: '온라인', color: 'secondary', style: 'weak' };
     case 'TEAM':
       return { text: '팀', color: 'secondary', style: 'fill' };
+    default:
+      return { text: '', color: 'secondary', style: 'fill' };
   }
 };
